% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancersea.R
\docType{data}
\name{available_pathways}
\alias{available_pathways}
\title{Available cancer-related pathways in this package}
\format{An object of class \code{character} of length 14.}
\usage{
available_pathways
}
\description{
Available cancer-related pathways in this package
}
\keyword{datasets}
