% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cancersea.R
\docType{data}
\name{angiogenesis}
\alias{angiogenesis}
\alias{apoptosis}
\alias{cell_cycle}
\alias{differentiation}
\alias{dna_damage}
\alias{dna_repair}
\alias{emt}
\alias{hypoxia}
\alias{inflammation}
\alias{invasion}
\alias{metastasis}
\alias{proliferation}
\alias{quiescence}
\alias{stemness}
\title{Curated CancerSEA gene sets}
\format{A tibble with columns ensembl_gene_id and symbol}
\source{
\url{http://biocc.hrbmu.edu.cn/CancerSEA/goDownload}
}
\usage{
angiogenesis

apoptosis

cell_cycle

differentiation

dna_damage

dna_repair

emt

hypoxia

inflammation

invasion

metastasis

proliferation

quiescence

stemness
}
\description{
To load, call \code{data(pathway_name)}.
}
\keyword{datasets}
