% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corLinkMat.R
\name{corLinkMat}
\alias{corLinkMat}
\title{Function: Correlation between spatial link matrices}
\usage{
corLinkMat(A, B)
}
\arguments{
\item{A}{First coded spatial link matrix}

\item{B}{Second coded spatial link matrix}
}
\value{
Scalar value correlation between A and B
}
\description{
\code{corLinkMat} evaluates the correlation between two link matrices
}
\details{
The function \code{corLinkMat(A, B)} evaluates the correlation between two
  Moran's \emph{I} spatial autocorrelation coefficients that are defined by
  different spatial link matrices or one link matrix in different coding schemes.
  The underlying assumptions are [a] that Moran's \emph{I} is based on a simple
  projection matrix \eqn{ M_[1]=I-1*(1^T*1)^-1*1^T} and [b] global spatial
  independence. The function transforms both matrices to symmetry
  by \eqn{(A+A^T)/2}.
}
\examples{
library(CancerSEA)
data(mig)
G <- mig[ ,-c(1:8)]                              # Extract migration matrix
n <- nrow(G)

Gmin <- matrix(0,n,n)                            # minimum flow between i and j
Gmax <- matrix(0,n,n)                            # maximum flow between i and j
for (i in 1:(n-1))
   for (j in (i+1):n){
      Gmin[i,j] <- Gmin[j,i] <- min(G[i,j],G[j,i])
      Gmax[i,j] <- Gmax[j,i] <- max(G[i,j],G[j,i])
   }

Cmin <- n/sum(Gmin)*Gmin                         # Global C-coding scheme
Cmax <- n/sum(Gmax)*Gmax                         # Global C-coding scheme

Wmin <- Gmin/rowSums(Gmin)                       # Row-sum W-coding scheme
Wmax <- Gmax/rowSums(Gmax)                       # Row-sum W-coding scheme

## Spectrum of eigenvalues
eval <- sort(eigen(Wmin, only.values = TRUE)$values)
plot(eval, type="h", main="Eigenvalue Spectrum of Wmin"); abline(h=0,lwd=2)

## Correlation between minimum and maximum link matrices
corLinkMat(Cmin,Cmax)                            # norun: 0.8933488
corLinkMat(Wmin,Cmin)                            # norum: 0.6848093

}
\references{
Michael Tiefelsdorf (2000). Modelling Spatial Processes. The
  Identification and Analysis of Spatial Relationships in Regression
  Residuals by Means of Moran's \emph{I}. Berlin: Springer
}
\author{
Michael Tiefelsdorf <tiefelsdorf@utd.edu>
}
