% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapColorQual.R
\name{mapColorQual}
\alias{mapColorQual}
\title{Function: Maps a qualitative theme of maximum 12 categories}
\usage{
mapColorQual(var.name, shape, map.title = "",
  legend.title = deparse(substitute(var.name)), legend.pos = "bottomleft",
  legend.cex = 1, add.to.map = F)
}
\arguments{
\item{var.name}{A factor variable to be mapped with a maximum of 12 categories.
If it is in a data-frame, then the data-frame must be refered to, e.g., \code{df$var}}

\item{shape}{An existing spatial polygon or spatial polygon data-frame}

\item{map.title}{Character string with map title}

\item{legend.title}{Character string with legend title (default=\code{var.name})}

\item{legend.pos}{Location of legend in the map frame (default=\code{"bottomleft"})}

\item{legend.cex}{Relative font size of the legend}

\item{add.to.map}{Logical to start a new map frame if \code{FALSE} or overlay onto an
existing map frame if \code{TRUE}}
}
\value{
\code{NULL}
}
\description{
\code{mapColorQual} generates a map of a qualitative variable
}
\details{
The function \code{mapColorQual} maps a \emph{categorical variable} with a
  \emph{set of distinct colors}. A legend is generated.
}
\examples{
library(maptools)
data(shpSEA); data(shpNA); data(shpStates)

## Generate map
boxSEA <- bbox(shpSEA)
shpSEA$URBRURFAC <- factor(shpSEA$URBRUR, labels=c("rural","urban"))
plot(shpNA, col="cornsilk", axes=T, xlim=boxSEA[1,], ylim=boxSEA[2,], bg="lightskyblue1")
mapColorQual(shpSEA$URBRURFAC, shpSEA, legend.cex=0.9, legend.pos="bottomright",
             legend.title="Urban/Rural",
             map.title="SEAs Classified as Urban and Rural", add.to.map=T)
plot(shpStates, border="red", lwd=2, add=TRUE)

}
\author{
Michael Tiefelsdorf <tiefelsdorf@utd.edu>
}
