% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva.R
\name{state_score}
\alias{state_score}
\title{The functional state scores of single cells}
\usage{
state_score(
  expr,
  gene_set,
  min_size = 1,
  max_size = Inf,
  parallel_size = 1L,
  verbose = FALSE
)
}
\arguments{
\item{expr}{Gene expression data which can be given as a matrix of expression values where rows correspond to genes and columns correspond to cells or samples.}

\item{gene_set}{Gene sets provided as a list, it must be match with the gene in expression profiles.}

\item{min_size}{Minimum size of the resulting gene sets.}

\item{max_size}{Maximum size of the resulting gene sets.}

\item{parallel_size}{Number of threads of execution to use when doing the calculations in parallel.}

\item{verbose}{Gives information about each calculation step. Default: FALSE.}
}
\description{
The functional state scores of single cells
}
\details{
The functional state scores of single cells
}
\author{
Erjie Zhao
}
