% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.lmHetero.R
\name{summary.lmHetero}
\alias{summary.lmHetero}
\title{Function: Summerize lmHetero model estimates}
\usage{
summary(object)
}
\arguments{
\item{object}{An object of class \code{lmHetero}.}
}
\value{
the input object is returned silently.
}
\description{
\code{summary.lmHetero} provides relevant summary information of a
heteroscedastic regression model estimated by \code{lmHetero}.
}
\details{
Regression tables for the estimate regression coefficients and the
coefficients of the weights model are provided as well as a maximum
likelihood ratio test is provided agaist a model, which assumes homoscedasticity.
}
\examples{
library(CancerSEA)
mod.lmH <- lmHetero(B_WM_P2_RT~L_WM_P1_RT+RAD_MD+URBRUR|log(POPATRISK1982), data=cancer)
summary(mod.lmH)


}
\author{
Michael Tiefelsdorf \email{tiefelsdorf@utd.edu}
}
