% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{angiogenesis}
\alias{angiogenesis}
\alias{apoptosis}
\alias{cell_cycle}
\alias{differentiation}
\alias{dna_damage}
\alias{dna_repair}
\alias{emt}
\alias{hypoxia}
\alias{inflammation}
\alias{invasion}
\alias{metastasis}
\alias{proliferation}
\alias{quiescence}
\alias{stemness}
\title{Curated CancerSEA gene sets}
\format{
A tibble with columns ensembl_gene_id and symbol

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 66 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 137 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 201 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 110 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 119 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 90 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 83 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 112 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 97 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 166 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 88 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 66 rows and 2 columns.

An object of class \code{spec_tbl_df} (inherits from \code{tbl_df}, \code{tbl}, \code{data.frame}) with 166 rows and 2 columns.
}
\source{
\url{http://biocc.hrbmu.edu.cn/CancerSEA/goDownload}
}
\usage{
angiogenesis

apoptosis

cell_cycle

differentiation

dna_damage

dna_repair

emt

hypoxia

inflammation

invasion

metastasis

proliferation

quiescence

stemness
}
\description{
To load, call \code{data(pathway_name)}.
}
\keyword{datasets}
